requirements.txt
pandas~=2.2.2
kaggle~=1.6.12
psycopg2~=2.9.9
SQLAlchemy~=2.0.29
python-dotenv~=1.0.1